backend/ingestion_service/pyproject.toml
[tool.poetry]
name = "ingestion-service"
version = "0.1.0"
description = "Repo ingestion microservice (FastAPI + Qdrant + Postgres)"
authors = ["ExyKK <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }

pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"

sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
psycopg = { version = "^3.2.1", extras = ["binary"] }

gitpython = "^3.1.43"
qdrant-client = "^1.11.0"

numpy = "^2.0.0"
torch = "^2.4.0"
transformers = "^4.44.0"
sentence-transformers = "^3.0.1"

tree-sitter = "0.25.2"
tree-sitter-go = "^0.25.0"
tree-sitter-html = "^0.23.0"
tree-sitter-json = "^0.21.0"
tree-sitter-yaml = "^0.6.0"
tree-sitter-toml = "^0.6.0"
tree-sitter-markdown = "^0.3.2"
tree-sitter-css = "^0.21.0"
tree-sitter-dockerfile = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
httpx = "^0.27.0"
ruff = "^0.6.0"
mypy = "^1.11.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
